gallery revised and updates to columns
changed all the images and added new larger ones or thumbnails.
changed headings to h3 instead of h4
Fixed the columns finally, it was the flex needing to be 25% so there are 2 and evenly spaced with 5% padding.
added back the dotted border and the hover /enlarge feature on the images.